refactor compiler to be pluggable
